chore(*): Move ConstraintEvaluator to plugins package
